// types/global.d.ts
export type SocialPosts = {
  uid?: string;
  username: string;
  email: string;
  profileImage: StaticImageData | string;
  postImage?: StaticImageData | string;
  type: {
    image: string;
    video: string;
  };
  video?: string;
  content: string;
  likes?: number;
  comments?: string;
  time:Firebase.Timestamp; 
//TODO change to Date.now or something similar
}
// ...
export type ProfileData = {
  firstName: string;
  lastName: string;
  banner: string;
  email: string;
  country: string;
  phone: string;
  aboutMe: string;
  likes: number;
  posts: number;
  following: number
  followers: number;
  birthday: Date | Firebase.Timestamp;
  school: string;
  workingAt: string;
  followerCount: number;
  followingCount: number;
}

export type UserData = {
	email: string;
	friends: string[];
	gender: string;
  isAdmin: boolean;
  last_loggedIn: Firebase.Timestamp;
  userName: string;
  photoURL: string;
};

export type Communities = {
  id: string;
  name: string;
  communityImage: StaticImageData | string;
  memberCount: number; // Renamed members to memberCount to hold the number of members
  members: string[]; // New field for storing the members' IDs or names
  category: string;
  isPublic: boolean;
  posts: SocialPostsData[];
}

export type Follower = {
  uid: string;
  username: string;
  profilePicture: string;
  email: string;
};

export type Following = {
  uid: string;
  username: string;
  profilePicture: string;
  email: string;
};